Refactoring respository attempting to split off generic ("shared") code from onehop test specific code. Adding some "navigational" README files in various packages.
