+++ tests/README.md
@@ -1,0 +1,5 @@
+# Tests
+
+The current SRI test suites currently implemented are:
+
+- ["One Hop" Tests for KPs and ARAs](tests/onehop/README.md)
