+++ translator/README.md
@@ -1,0 +1,7 @@
+# Shared Translator Modules
+
+This package contains various utility modules supporting specific subtypes of SRI Testing:
+
+- [translator.sri.testing](sri/testing): utility methods for initializing globally shared parameters (default Biolink Model and TRAPI versions used for tests) and for creating test templates.
+  - [translator.sri.testing.util](sri/testing/util): modules for ontology knowledge provider interfacing and to report missing predicates.
+- [translator.trapi](trapi): Translator Reasoner Application Programming Interface interfacing code.
