+++ translator/sri/testing/README.md
@@ -1,0 +1,5 @@
+# SRI Testing Package
+
+- [Jupyter Notebook to facilitate access to ARA and KP TRAPI entries](SmartAPI.ipynb) in the Translator SmartAPI Registry.
+- Method ([`create_template.py`](create_templates.py)) to create testing templates by interrogation of the Translator SmartAPI Registry.
+- [Utility Package](util).
