+++ translator/sri/testing/util/README.md
@@ -1,0 +1,4 @@
+# SRI Testing Utility Package methods
+
+- Ontology knowledge provider interfacing module
+- Module to run to compile and report missing predicates
