+++ translator/sri/testing/util/ontology_kp.py
@@ -5,6 +5,7 @@ import requests
 
 ONTOLOGY_KP_TRAPI_SERVER = "https://ontology-kp.apps.renci.org/query"
 
+
 def post(url, message, params=None):
     """
     :param url
