+++ translator/trapi/__init__.py
@@ -1,5 +1,5 @@
-from json import dumps
 from typing import Optional
+from json import dumps
 
 import logging
 
@@ -11,16 +11,20 @@ from reasoner_validator import validate
 logger = logging.getLogger(__name__)
 logger.setLevel("DEBUG")
 
-_default_trapi_version = None
+# TODO: We'd rather NOT hard code a default TRAPI here,
+#       but do it for now pending clarity on how to guide
+#       the choice of TRAPI from a Translator SmartAPI entry
+# Default is actually specifically 1.2.0 as of March 2022,
+# but the ReasonerAPI should discern this
+DEFAULT_TRAPI_VERSION = "1"
 
-# TODO: We'd rather NOT hard code a default TRAPI here, but do it for now pending clarity on how to guide
-#       the choice of TRAPI from the Translator SmartAPI for a given resource
-DEFAULT_TRAPI_VERSION = "1"  # actually specifically 1.2.0 as of March 2022, but the ReasonerAPI should discern this
+_default_trapi_version = None
 
 
 def set_trapi_version(version: str):
     global _default_trapi_version
     _default_trapi_version = version if version else DEFAULT_TRAPI_VERSION
+    logger.debug(f"TRAPI Version set to {_default_trapi_version}")
 
 
 def get_trapi_version() -> Optional[str]:
@@ -74,7 +78,7 @@ def call_trapi(url, opts, trapi_message):
     return {'status_code': response.status_code, 'response_json': response_json}
 
 
-def output(json):
+def _output(json):
     return dumps(json, sort_keys=False, indent=4)
 
 
@@ -92,7 +96,7 @@ def execute_trapi_lookup(case, creator, rbag):
     if trapi_request is None:
         # The particular creator cannot make a valid message from this triple
         assert False, f"\nCreator method '{creator.__name__}' for test case \n" + \
-                      f"\t{output(case)}\n" + \
+                      f"\t{_output(case)}\n" + \
                       f"could not generate a valid TRAPI query request object?"
 
     # query use cases pertain to a particular TRAPI version
@@ -101,7 +105,7 @@ def execute_trapi_lookup(case, creator, rbag):
     if not is_valid_trapi(trapi_request, trapi_version=trapi_version):
         # This is a problem with the testing framework.
         assert False, f"execute_trapi_lookup({case['url']}): Invalid TRAPI '{trapi_version}' " + \
-                      f"query request {output(trapi_request)}"
+                      f"query request {_output(trapi_request)}"
 
     trapi_response = call_trapi(case['url'], case['query_opts'], trapi_request)
 
@@ -111,33 +115,33 @@ def execute_trapi_lookup(case, creator, rbag):
 
     if trapi_response['status_code'] != 200:
         err_msg = f"execute_trapi_lookup({case['url']}): " + \
-                  f"TRAPI request:\n\t{output(trapi_request)}\n " + \
+                  f"TRAPI request:\n\t{_output(trapi_request)}\n " + \
                   f"returned status code: {str(trapi_response['status_code'])} " + \
-                  f"and response:\n\t '{output(trapi_response['response_json'])}'"
+                  f"and response:\n\t '{_output(trapi_response['response_json'])}'"
         logger.warning(err_msg)
         assert False, err_msg
 
     # Validate that we got back valid TRAPI Response
     assert is_valid_trapi(trapi_response['response_json'], trapi_version=trapi_version), \
            f"execute_trapi_lookup({case['url']}): " + \
-           f"TRAPI request:\n\t{output(trapi_request)}\n " + \
+           f"TRAPI request:\n\t{_output(trapi_request)}\n " + \
            f"had an invalid TRAPI '{trapi_version}' response:\n\t" + \
-           f"{output(trapi_response['response_json'])}"
+           f"{_output(trapi_response['response_json'])}"
 
     response_message = trapi_response['response_json']['message']
 
     # Verify that the response had some results
     assert len(response_message['results']) > 0, \
            f"execute_trapi_lookup({case['url']}): empty TRAPI Result from TRAPI request:\n\t" + \
-           f"{output(trapi_request)}"
+           f"{_output(trapi_request)}"
 
     # The results contained the object of the query
     object_ids = [r['node_bindings'][output_node_binding][0]['id'] for r in response_message['results']]
     assert case[output_element] in object_ids, \
            f"execute_trapi_lookup({case['url']}): " + \
-           f"TRAPI request:\n\t{output(trapi_request)}\n " + \
+           f"TRAPI request:\n\t{_output(trapi_request)}\n " + \
            f"had missing or invalid TRAPI Result object ID bindings" \
            f"in response method results:\n\t" \
-           f"{output(response_message['results'])}"
+           f"{_output(response_message['results'])}"
 
     return response_message
